+++ StripePaymentSheet/StripePaymentSheet/Source/Helpers/PaymentSheetLinkAccount.swift
@@ -62,6 +62,9 @@ class PaymentSheetLinkAccount: PaymentSheetLinkAccountInfoProtocol {
 
     var paymentSheetLinkAccountDelegate: PaymentSheetLinkAccountDelegate?
 
+    var phoneNumberUsedInSignup: String?
+    var nameUsedInSignup: String?
+
     let email: String
 
     var redactedPhoneNumber: String? {
@@ -371,6 +374,7 @@ class PaymentSheetLinkAccount: PaymentSheetLinkAccountInfoProtocol {
         cvc: String?,
         allowRedisplay: STPPaymentMethodAllowRedisplay?,
         expectedPaymentMethodType: String?,
+        billingPhoneNumber: String?,
         completion: @escaping (Result<PaymentDetailsShareResponse, Error>
     ) -> Void) {
         retryingOnAuthError(completion: completion) { [apiClient, publishableKey] completionRetryingOnAuthErrors in
@@ -389,6 +393,7 @@ class PaymentSheetLinkAccount: PaymentSheetLinkAccountInfoProtocol {
                 cvc: cvc,
                 allowRedisplay: allowRedisplay,
                 expectedPaymentMethodType: expectedPaymentMethodType,
+                billingPhoneNumber: billingPhoneNumber,
                 consumerAccountPublishableKey: publishableKey,
                 completion: completionRetryingOnAuthErrors
             )
@@ -492,8 +497,14 @@ extension PaymentSheetLinkAccount {
     /// Returns `nil` if not authenticated/logged in.
     ///
     /// - Parameter paymentDetails: Payment details
+    /// - Parameter cvc: The CVC that we need to pass for some transactions
+    /// - Parameter billingPhoneNumber: The billing phone number to add to the params. Passing it separately because it's not part of the payment details.
     /// - Returns: Payment method params for paying with Link.
-    func makePaymentMethodParams(from paymentDetails: ConsumerPaymentDetails, cvc: String?) -> STPPaymentMethodParams? {
+    func makePaymentMethodParams(
+        from paymentDetails: ConsumerPaymentDetails,
+        cvc: String?,
+        billingPhoneNumber: String?
+    ) -> STPPaymentMethodParams? {
         guard let currentSession = currentSession else {
             stpAssertionFailure("Cannot make payment method params without an active session.")
             return nil
@@ -501,6 +512,7 @@ extension PaymentSheetLinkAccount {
 
         let params = STPPaymentMethodParams(type: .link)
         params.billingDetails = STPPaymentMethodBillingDetails(billingAddress: paymentDetails.billingAddress, email: paymentDetails.billingEmailAddress)
+        params.billingDetails?.phone = billingPhoneNumber
         params.link?.paymentDetailsID = paymentDetails.stripeID
         params.link?.credentials = ["consumer_session_client_secret": currentSession.clientSecret]
 
@@ -594,7 +606,7 @@ private extension LinkSettings.FundingSource {
 
 struct UpdatePaymentDetailsParams {
     enum DetailsType {
-        case card(expiryDate: CardExpiryDate, billingDetails: STPPaymentMethodBillingDetails? = nil)
+        case card(expiryDate: CardExpiryDate? = nil, billingDetails: STPPaymentMethodBillingDetails? = nil)
         // updating bank not supported
     }