+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Link/PaymentSheet-LinkConfirmOption.swift
@@ -33,7 +33,8 @@ extension PaymentSheet {
         /// Confirm intent with paymentDetails.
         case withPaymentDetails(
             account: PaymentSheetLinkAccount,
-            paymentDetails: ConsumerPaymentDetails
+            paymentDetails: ConsumerPaymentDetails,
+            confirmationExtras: LinkConfirmationExtras?
         )
     }
 
@@ -51,7 +52,7 @@ extension PaymentSheet.LinkConfirmOption {
             return account
         case .withPaymentMethod:
             return nil
-        case .withPaymentDetails(let account, _):
+        case .withPaymentDetails(let account, _, _):
             return account
         }
     }
@@ -64,8 +65,17 @@ extension PaymentSheet.LinkConfirmOption {
             return intentConfirmParams.paymentMethodParams.paymentSheetLabel
         case .withPaymentMethod(let paymentMethod):
             return paymentMethod.paymentSheetLabel
-        case .withPaymentDetails(_, let paymentDetails):
+        case .withPaymentDetails(_, let paymentDetails, _):
             return paymentDetails.paymentSheetLabel
+        }
+    }
+
+    var paymentMethodType: String {
+        switch self {
+        case .signUp(_, _, _, _, let intentConfirmParams):
+            return intentConfirmParams.paymentMethodParams.type.identifier
+        case .wallet, .withPaymentMethod, .withPaymentDetails:
+            return STPPaymentMethodType.link.identifier
         }
     }
 
@@ -77,7 +87,7 @@ extension PaymentSheet.LinkConfirmOption {
             return intentConfirmParams.paymentMethodParams.billingDetails
         case .withPaymentMethod(let paymentMethod):
             return paymentMethod.billingDetails
-        case .withPaymentDetails(_, let paymentDetails):
+        case .withPaymentDetails(_, let paymentDetails, _):
             return STPPaymentMethodBillingDetails(billingAddress: paymentDetails.billingAddress, email: paymentDetails.billingEmailAddress)
         }
     }