+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 24.11.0 2025-04-07
+### PaymentSheet
+* [Added] Added `rowSelectionBehavior` API to Embedded Payment Element (private preview).
+* [Changed] The native Link experience (private preview) is now available as a payment method even if you collect billing details via `billingDetailsCollectionConfiguration`.
+* [Changed] Updated Afterpay branding in the US to be Cash App Afterpay.
+
+## 24.10.0 2025-03-31
+### PaymentSheet
+* [Added] Added `LinkConfiguration` to allow control over Link in PaymentSheet.
+* [Fixed] Improved reliability of expiration date detection when scanning cards.
+
 ## 24.9.0 2025-03-24
 ### Connect (Private Preview)
 * [Fixed] Fixed an issue where Account Onboarding exit events were not being properly emitted.